// frontend/src/app/shared/model/EmployeeDetails.ts
export class EmployeeDetails {
  userName: string;
  age: number;
  email: string;
  accNo: string;
  firstName: string;
  lastName: string;
  ifsc: string;
  nextPayDate: string;
  dob: string;
  phone: string;
  salary: number;
  taxPercent: number;
  bonus: number;
  excemptionAmt: number;
  payFreq: number;
  overTime: number;
  role:string;
}

export class AthTableData{
  userName:string;
  password:string;
  role:number;
}